feat: move legacy bridge to a separate page
