+++ apps/web/src/app/Routes.tsx
@@ -21,6 +21,7 @@ const TradePageLayout = lazyLoad(
 import { TradePage } from './pages/trade/page';
 import { SupplyPage } from './pages/trade/supply/page';
 import { BridgePage } from './pages/trade/bridge/page';
+import { LegacyBridge } from './pages/legacy-bridge/page';
 
 const VotePage = lazyLoad(
   () => import('./pages/vote/page'),
@@ -95,6 +96,7 @@ export default function RootRoutes() {
           <Route path="proposal/:id" element={<ProposalDetailsPage />} />
         </Route>
 
+        <Route path="legacy-bridge" element={<LegacyBridge />} />
         <Route path="airdrip" element={<Redirect to="https://balanced.network/" />} />
         <Route path="*" element={<Redirect to="https://balanced.network/404" />} />
       </Route>
