+++ apps/web/src/app/pages/trade/layout.tsx
@@ -5,42 +5,16 @@ import { useIconReact } from 'packages/icon-react';
 import { Outlet, useLocation, useNavigate } from 'react-router-dom';
 
 import { Flex, Box } from 'rebass/styled-components';
-import styled, { css } from 'styled-components';
 
 import { useArchwayContext } from 'app/_xcall/archway/ArchwayProvider';
 import { UnderlineText } from 'app/components/DropdownText';
 import { Tab, Tabs } from 'app/components/Tab';
-import { useTransferAssetsModalToggle } from 'store/application/hooks';
 import { useFetchBBalnInfo, useFetchBBalnSources } from 'store/bbaln/hooks';
 import { useFetchOraclePrices } from 'store/oracle/hooks';
 import { useFetchPrice } from 'store/ratio/hooks';
 import { useFetchRewardsInfo } from 'store/reward/hooks';
 import { useFetchStabilityFundBalances } from 'store/stabilityFund/hooks';
 import { useWalletFetchBalances } from 'store/wallet/hooks';
-
-const BTPButton = styled(UnderlineText)`
-  padding-right: 0 !important;
-  font-size: 14px;
-  padding-bottom: 5px;
-  margin: 5px 0 20px;
-  width: 250px;
-  display: none;
-
-  ${({ theme }) => theme.mediaWidth.upSmall`
-    position: absolute;
-    align-self: flex-end;
-    transform: translate3d(0, 9px, 0);
-    padding-bottom: 0;
-    margin: 0;
-    width: auto;
-    display: block;
-  `};
-
-  ${({ theme }) =>
-    css`
-      color: ${theme.colors.primaryBright};
-    `};
-`;
 
 export function TradePageLayout() {
   const { account } = useIconReact();
@@ -76,17 +50,10 @@ export function TradePageLayout() {
     }
   };
 
-  //handle wallet modal
-  const toggleTransferAssetsModal = useTransferAssetsModalToggle();
-  const handleBTPButtonClick = () => {
-    toggleTransferAssetsModal();
-  };
-
   return (
     <>
       <Box flex={1}>
         <Flex mb={10} flexDirection="column">
-          <BTPButton onClick={handleBTPButtonClick}>Transfer assets between blockchains (Legacy)</BTPButton>
           <Flex alignItems="center" justifyContent="space-between">
             <Tabs value={value} onChange={handleTabClick}>
               <Tab>
